requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
minio==7.1.14
pandas
SQLAlchemy
psycopg2-binary
apache-airflow-providers-docker
crypto
apache-airflow-providers-google
pydrive2